% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ragg-package.R
\docType{package}
\name{ragg-package}
\alias{ragg}
\alias{ragg-package}
\title{ragg: Graphic Devices Based on AGG}
\description{
Anti-Grain Geometry (AGG) is a high-quality and high-performance
    2D drawing library. The 'ragg' package provides a set of graphic devices 
    based on AGG to use as alternative to the raster devices provided through
    the 'grDevices' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomasp85/ragg}
  \item Report bugs at \url{https://github.com/thomasp85/ragg/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomas.pedersen@rstudio.com} (0000-0002-5147-4711)

Authors:
\itemize{
  \item Maxim Shemanarev (Author of AGG)
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
