% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{register_font}
\alias{register_variant}
\alias{font_feature}
\alias{get_font_features}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{systemfonts}{\code{\link[systemfonts]{font_feature}}, \code{\link[systemfonts]{register_font}}, \code{\link[systemfonts]{register_variant}}}

  \item{textshaping}{\code{\link[textshaping]{get_font_features}}}
}}

