% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ragg-package.R
\docType{package}
\name{ragg-package}
\alias{ragg}
\alias{ragg-package}
\title{ragg: Graphic Devices Based on AGG}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Anti-Grain Geometry (AGG) is a high-quality and high-performance 2D drawing library. The 'ragg' package provides a set of graphic devices based on AGG to use as alternative to the raster devices provided through the 'grDevices' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ragg.r-lib.org}
  \item \url{https://github.com/r-lib/ragg}
  \item Report bugs at \url{https://github.com/r-lib/ragg/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomas.pedersen@posit.co} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

Authors:
\itemize{
  \item Maxim Shemanarev (Author of AGG) [copyright holder]
}

Other contributors:
\itemize{
  \item Tony Juricic \email{tonygeek@yahoo.com} (Contributor to AGG) [contributor, copyright holder]
  \item Milan Marusinec \email{milan@marusinec.sk} (Contributor to AGG) [contributor, copyright holder]
  \item Spencer Garrett (Contributor to AGG) [contributor]
  \item Posit, PBC [copyright holder, funder]
}

}
\keyword{internal}
