% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{register_font}
\alias{register_variant}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{systemfonts}{\code{\link[systemfonts]{register_font}}, \code{\link[systemfonts]{register_variant}}}
}}

